Update: Certifications methods now extends ResourceBase
